requirements.txt
asgiref==3.2.10
awsebcli==3.20.10
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.3
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-session-timeout==0.1.0
idna==3.7
jmespath==1.0.1
pathspec==0.10.1
pillow==10.2.0
psycopg2==2.9.9
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
